requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.6
django-ckeditor==6.7.1
django-import-export==4.0.3
django-js-asset==2.2.0
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.11.0
whitenoise==6.6.0
